6 - Add health check Added a plug that checks if request is a health check, i.e. is for /health. Returns a 200 in yes and passes the through if not.
A health check is needed to allow any monitoring tool to detect
if app is healthy. Note the need to add a health check only arose
when a load balancer was needed to route incoming traffic to distributed
app.

The implemented solution was copied from https://blog.jola.dev/health-checks-for-plug-and-phoenix