33 Add download endpoint
Added download endpoint and new getter that returns decks with the cards
included in the response.It is probably a good idea to simply use the
list_decks getter that takes an expand query params.

Also added jason encoder rule for the decks and updated the ones of card and
answer entities.