49 Restrict decks access by user id and include skipped tests
The access to the decks is restriced by user id.Took the opportunity to
change the API of the decks so that all methods accept a map. This
essentially makes a the decks domain a black box with which one
communicates in a very consistent manner.

Included the skipped tests and where necessary updated the returns
value decks so that the cards field is always Ecto.NotLoaded.